python-backend/pyproject.toml
[tool.poetry]
name = "food-vision-backend"
version = "0.1.0"
description = ""
authors = ["Matt Sárdi <sardi.matt@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.14.2"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
pyprojroot = "^0.3.0"
transformers = "^4.38.2"
scikit-image = "^0.22.0"
pillow = "^10.2.0"
torch = "^2.2.1"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
python-multipart = "^0.0.9"


[tool.poetry.group.dev.dependencies]
matplotlib = "^3.8.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
